+++ src/Components/Core/Presentation/React/LearningSpaceDisplay/LearningElementModal/ILearningElementModalController.ts
@@ -1,7 +1,7 @@
 export default interface ILearningElementModalController {
   closeModal(): void;
   triggerOutroCutscene(): void;
-  scoreLearningElement(): Promise<void>;
+  scoreLearningElement(): Promise<boolean>;
   h5pEventCalled(event: any): Promise<void>;
   xAPICompletedListener(t: any): void;
   showBottomToolTip(): void;